refactor(web): extract socket and tts hooks
